+++ nomad/mock/network.go
@@ -17,3 +17,15 @@ func NewNetworkStatus(address string) structs.NetworkStatus {
 func (ns *NetworkStatus) NetworkStatus() *structs.AllocNetworkStatus {
 	return &structs.AllocNetworkStatus{Address: ns.address}
 }
+
+func AllocNetworkStatus() *structs.AllocNetworkStatus {
+	return &structs.AllocNetworkStatus{
+		InterfaceName: "eth0",
+		Address:       "192.168.0.100",
+		DNS: &structs.DNSConfig{
+			Servers:  []string{"1.1.1.1"},
+			Searches: []string{"localdomain"},
+			Options:  []string{"ndots:5"},
+		},
+	}
+}
